Sync (#5)
* Strip all whitespace from placeholders in RecipientCSV

We were not consistently stripping whitespace from around placeholders
in the `RecipientCSV` class. We were left stripping all placeholders
because of the `skipinitialspace=True` argument we pass to `csv.reader`
and we were also right stripping the last placeholder in the file, but
no other white space was removed.

In notifications-admin the extra whitespace doesn't appear (because
multiple white space characters only appear as one whitespace charcter
in HTML), which means that what got sent and what the user sees in admin
are different, and the `billable_units` can be confusing.

* Bump version to 33.2.0 and upgrade a couple of dependencies

* Put space between r and slash in <br /> tag

Our emails use the XHTML 1.0 strict doctype. The XHTML 1.0 strict
specification recommends having a space between the `br` and the closing
`/` of the `<br />` tag, for compatibility with existing user agents[1],
 although only the self-closing slash is required[2].

Some versions of the Alpine Mail client only recognise the tag with the
space before the slash[3]. Users of this client see multiple lines
concatenated into one in our emails, when the author of the email has
used a single linebreak.

1. https://www.w3.org/TR/xhtml1/#C_2
2. https://www.w3.org/TR/xhtml1/#h-4.6
3. http://mailman13.u.washington.edu/pipermail/alpine-info/2014-March/005718.html