Add `get_total_hard_bounces` to the bounce rate client (#208)
This PR adds the get_total_hard_bounces method to our bounce rate client. This is so we can populate this number on the dashboard from Redis and will be used over in this PR: cds-snc/notification-admin#1554

I also added get_total_notifications in case we need that later, and some docstrings.